--- next/adapter/postgres/doc.go
@@ -1,2 +1,0 @@
-// Package postgres contains gateways for postgres SQL
-package postgres
