+++ next/adapter/sqlite/doc.go
@@ -1,0 +1,2 @@
+// Package sqlite contains gateways for sqlite SQL
+package sqlite
